+++ packages/amplify_api/test/auth_providers_test.mocks.dart
@@ -35,4 +35,6 @@ class MockOIDCAuthProvider extends _i1.Mock implements _i2.OIDCAuthProvider {
   _i4.Future<String?> getLatestAuthToken() =>
       (super.noSuchMethod(Invocation.method(#getLatestAuthToken, []),
           returnValue: Future<String?>.value()) as _i4.Future<String?>);
+  @override
+  String toString() => super.toString();
 }