+++ packages/amplify_auth_cognito/android/src/main/kotlin/com/amazonaws/amplify/amplify_auth_cognito/AuthCognito.kt
@@ -16,6 +16,7 @@
 package com.amazonaws.amplify.amplify_auth_cognito
 
 import android.app.Activity
+import android.app.Activity.RESULT_CANCELED
 import android.content.Context
 import android.content.Intent
 import android.os.Handler
@@ -53,6 +54,7 @@ import com.amazonaws.amplify.amplify_auth_cognito.types.FlutterSignOutRequest
 import com.amazonaws.amplify.amplify_core.exception.ExceptionUtil.Companion.handleAddPluginException
 import com.amazonaws.amplify.amplify_auth_cognito.utils.isRedirectActivityDeclared
 import com.amazonaws.mobile.client.AWSMobileClient
+import com.amazonaws.amplify.amplify_core.AtomicResult
 import com.amplifyframework.auth.AuthException
 import com.amplifyframework.auth.AuthProvider
 import com.amplifyframework.auth.AuthSession
@@ -131,9 +133,14 @@ public class AuthCognito : FlutterPlugin, ActivityAware, MethodCallHandler, Plug
 
   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?): Boolean {
     var isHostedUIActivity = isRedirectActivityDeclared(context)
-    if (!isHostedUIActivity && requestCode == AWSCognitoAuthPlugin.WEB_UI_SIGN_IN_ACTIVITY_CODE) {
-      Amplify.Auth.handleWebUISignInResponse(data)
-      return true
+    if (requestCode == AWSCognitoAuthPlugin.WEB_UI_SIGN_IN_ACTIVITY_CODE) {
+        /// The HostedUI activity in amplify-android handles success case
+        /// We need a response handler if the HostedUI activity isn't used...
+         /// ... or if the HostedUI activity is used, but the user cancels
+        if (!isHostedUIActivity || resultCode == RESULT_CANCELED) {
+            Amplify.Auth.handleWebUISignInResponse(data)
+            return true
+        }
     }
     return false
   }
@@ -152,7 +159,9 @@ public class AuthCognito : FlutterPlugin, ActivityAware, MethodCallHandler, Plug
     return args as HashMap<String, Any>
   };
 
-  override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
+  override fun onMethodCall(@NonNull call: MethodCall, @NonNull _result: Result) {
+
+    val result = AtomicResult(_result, call.method)
 
     if(call.method == "addPlugin"){
       try {