+++ packages/amplify_auth_cognito/lib/amplify_auth_cognito_stream_controller.dart
@@ -53,10 +53,7 @@ _onListen() {
           }
           break;
         default:
-          {
-            print(
-                'An Unrecognized Auth Hub event has been detected on the event channel.');
-          }
+          break;
       }
     });
   }