+++ packages/amplify_datastore/CHANGELOG.md
@@ -43,6 +43,27 @@ amplify codegen models
 
 - break(datastore): cannot saving boolean as integer in SQLite (#895)
 
+## 0.2.9 (2021-11-17)
+
+- chore: upgrade amplify-android to 1.28.3-rc
+
+## 0.2.8 (2021-11-12)
+
+### Fixes
+
+- fix(datastore): (Android) Fix DataStore release mode crash (#1064)
+
+## 0.2.7 (2021-11-08)
+
+### Chores
+
+- chore: Bump Amplify iOS to 1.15.5
+
+### Fixes
+
+- fix(datastore): Temporal date/time query predicates
+- fix(datastore): Android TemporalTime Save Issue
+
 ## 0.2.6 (2021-10-25)
 
 ### Fixes